pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for What To Cook? app."
authors = ["Timur Kasimov <tkasimov@icloud.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pymongo = "^4.3.3"
fastapi = "^0.88.0"
python-dotenv = "^0.21.0"
uvicorn = "^0.20.0"
omegaconf = "^2.3.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.17.0"
mypy = "^0.991"
flake8-colors = "^0.1.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
force_alphabetical_sort_within_sections = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
force_grid_wrap = 2

[tool.mypy]
python_version = 3.9

exclude = "tests/*"

disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
allow_untyped_globals = true
allow_redefinition = true

follow_imports = "skip"
